pyproject.toml
[project]
name = "kartheek-bs-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "arize-phoenix>=13.19.2",
    "arize-phoenix-otel>=0.15.0",
    "deepagents>=0.4.12",
    "langchain-openai>=1.1.12",
    "openinference-instrumentation-langchain>=0.1.61",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "pre-commit>=4.5.1",
]
